requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
redis
python-dotenv
pytest
httpx
requests
